pyproject.toml
[project]
name = "gh-action-pulse"
version = "0.1.0"
description = "Get Health/Propose modifications of each GitHub Action in a repository"
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "Jean-Yves LENHOF", email = "jygithub@lenhof.eu.org" }]
dependencies = [
  "packaging>=26.1",
  "pygithub>=2.9.0",
]

[project.scripts]
gh-action-pualse = "gh_action_pulse.main:main"

[dependency-groups]
dev = [
  "coverage>=7.13.5",
  "faker>=40.15.0",
  "pylint>=4.0.5",
  "pytest>=9.0.3",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.1.0",
  "pytest-mock>=3.15.1",
  "pytest-unordered>=0.7.0",
  "pytest-xdist>=3.8.0",
  "ruff>=0.15.10",
  "ty>=0.0.29",
]

[build-system]
requires = ["uv_build>=0.11.5,<0.12.0"]
build-backend = "uv_build"

[tool.ty.src]
include = [
  "src",
  "tests",
]
